% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adg_get.R
\name{adg_get}
\alias{adg_get}
\title{adg_get}
\usage{
adg_get(data, my_break, weighting, ...)
}
\arguments{
\item{data}{the results from function of outlier remove}

\item{my_break}{the weight breaks, with only one break}

\item{weighting}{the weighting choosed to select visit data}

\item{...}{linear model to get adg}
}
\value{
the adg of each animals
}
\description{
adg_get
}
