% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_transform.R
\name{csv_tranform}
\alias{csv_tranform}
\title{csv_transform}
\usage{
csv_tranform(station_type, data)
}
\arguments{
\item{station_type}{nedap or fire stations}

\item{data}{all csv data of locations}
}
\value{
nedap or fire stations with same transform format data
}
\description{
csv_transform
}
