% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfi_plot.R
\name{dfi_plot}
\alias{dfi_plot}
\title{dfi_plot}
\usage{
dfi_plot(dfi_data, path_out)
}
\arguments{
\item{dfi_data}{the resluts from dfi_get}

\item{path_out}{the path to put results}
}
\value{
pictures of dfi
}
\description{
dfi_plot
}
\examples{
#temp5 <- dfi_get(origin_data = temp1, adg_data = temp4)
}
