% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated_responder_info.R
\name{duplicated_responder_infos}
\alias{duplicated_responder_infos}
\title{duplicated_responder_infos}
\usage{
duplicated_responder_infos(data, begin_date)
}
\arguments{
\item{data}{the results from function of outlier_found}

\item{begin_date}{the begin data of test}
}
\value{
the duplicated responder infomations
}
\description{
duplicated_responder_infos
}
\examples{
#duplicated_responder_infos(temp2, begin_date = "2015-07-01")
}
