% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_code.R
\name{error_type_freq}
\alias{error_type_freq}
\title{error_type_freq}
\usage{
error_type_freq(data)
}
\arguments{
\item{data}{transformed data from function of transform}
}
\value{
frequency of 16 error types
}
\description{
error_type_freq
}
