% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_code.R
\name{feed_intake_monitor}
\alias{feed_intake_monitor}
\title{feed_intake_monitor}
\usage{
feed_intake_monitor(data, station_type, path_out, ...)
}
\arguments{
\item{data}{all csv data of locations}

\item{station_type}{nedap or fire stations}

\item{path_out}{the path to receive results}

\item{...}{ggsave parameters}
}
\value{
pictures of each locations
}
\description{
feed_intake_monitor
}
