% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fid_monitor.R
\name{fid_monitor}
\alias{fid_monitor}
\title{fid_monitor}
\usage{
fid_monitor(data, station_type, path_out, ...)
}
\arguments{
\item{data}{all csv data of locations}

\item{station_type}{nedap or fire stations}

\item{path_out}{the path to receive results}

\item{...}{ggsave parameters}
}
\value{
pictures of each locations
}
\description{
fid_monitor
}
\examples{
#fid_monitor(data = test, station_type = "nedap", path_out = "", width =60, height = 500)
}
