% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_data.R
\docType{data}
\name{fire_data}
\alias{fire_data}
\title{Data of fire location data}
\format{
A tibble with 26959 rows and 10 variables:
\describe{
\item{Location}{int location recorded when each visit}
\item{Tag}{integer64 responder of each animal}
\item{Date}{chr date of each visit}
\item{Entry}{chr entry time in a visit}
\item{Exit}{chr exit time in a visit}
\item{\verb{Ent Wt}}{num feed weight of entry in a visit}
\item{\verb{Ext Wt}}{num feed weight of exit in a visit}
\item{Consumed}{num feed intake in a visit}
\item{Weight}{num weight recorded in each visit}
\item{\verb{Topup Amount}}{num  topup amount}
}
}
\source{
fire location data of GuangXi YangXiang company
}
\usage{
fire_data
}
\description{
Data of fire location data
}
\keyword{datasets}
