% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_curve.R
\name{growth_curve}
\alias{growth_curve}
\title{growth_curve}
\usage{
growth_curve(data, path_out, weighting)
}
\arguments{
\item{data}{results from function of outlier_found}

\item{path_out}{the path to put your results}

\item{weighting}{the weighting you choose to determine outliers}
}
\value{
pictures of growth curve
}
\description{
growth_curve
}
\examples{
#growth_curve(data = temp2, w = 0.5, path_out = "")
}
