% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_monitor.R
\name{location_monitor}
\alias{location_monitor}
\title{location_monitor}
\usage{
location_monitor(data, station_type, path_out)
}
\arguments{
\item{data}{all csv data of locations}

\item{station_type}{nedap or fire stations}

\item{path_out}{the path to receive results}
}
\value{
pictures of each locations
}
\description{
location_monitor
}
