% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nedap_data.R
\docType{data}
\name{nedap_data}
\alias{nedap_data}
\title{Data of nedap location data}
\format{
A data.table with 32992 rows and 9 variables:
\describe{
\item{animal_number}{int responder of each animal}
\item{life_number}{logi no use}
\item{responder}{nteger64 responder of each animal}
\item{locationi}{int location recorded when each visit}
\item{visit_time}{chr date time in each visit}
\item{duration}{int times each visit}
\item{state}{int ...}
\item{weight}{int weight recorded in each visit}
\item{feed_intake}{feed intake in each visit}
}
}
\source{
nedap location data of GuangXi YangXiang company
}
\usage{
nedap_data
}
\description{
Data of nedap location data
}
\keyword{datasets}
