% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_code.R
\name{outlier_found}
\alias{outlier_found}
\title{outlier_found}
\usage{
outlier_found(data, ...)
}
\arguments{
\item{data}{transformed data from function of transform}

\item{...}{the rubost model fomula}
}
\value{
the weighting of each visit
}
\description{
outlier_found
}
