% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_code.R
\name{outlier_remove}
\alias{outlier_remove}
\title{outlier_remove}
\usage{
outlier_remove(data, begin_date)
}
\arguments{
\item{data}{the results from function of outlier_found}

\item{begin_date}{the begin data of test}
}
\value{
the remaining visit data from oulier_found
}
\description{
outlier_remove
}
