% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_remove.R
\name{outliers_remove}
\alias{outliers_remove}
\title{outliers_remove}
\usage{
outliers_remove(data, begin_date)
}
\arguments{
\item{data}{the results from function of outlier_found}

\item{begin_date}{the begin data of test}
}
\value{
the remaining visit data from oulier_found
}
\description{
outliers_remove
}
\examples{
#temp3 <- outliers_remove(temp2, begin_date = "2015-07-01")
}
