% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_code.R
\name{ppt_monitor}
\alias{ppt_monitor}
\title{ppt_monitor}
\usage{
ppt_monitor(data, station_type, path_out)
}
\arguments{
\item{data}{all csv data of locations}

\item{station_type}{nedap or fire stations}

\item{path_out}{the path to receive results}
}
\value{
pictures of each locations
}
\description{
ppt_monitor
}
