% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removed_freq.R
\name{removed_freq}
\alias{removed_freq}
\title{removed_freq}
\usage{
removed_freq(data, begin_date)
}
\arguments{
\item{data}{the results from function of outlier_found}

\item{begin_date}{the begin data of test}
}
\value{
the remaining visit data from oulier_found
}
\description{
removed_freq
}
\examples{
#removed_freq(data = temp2, begin_date = "2015-07-01")
}
