% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_monitor.R
\name{station_monitor}
\alias{station_monitor}
\title{location_monitor}
\usage{
station_monitor(data, station_type, path_out)
}
\arguments{
\item{data}{all csv data of locations}

\item{station_type}{nedap or fire stations}

\item{path_out}{the path to receive results}
}
\value{
pictures of each locations
}
\description{
location_monitor
}
\examples{
#station_monitor(data = test, station_type = "nedap", path_out = "")
}
