% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_code.R
\name{tranform}
\alias{tranform}
\title{transform}
\usage{
tranform(station_type, data)
}
\arguments{
\item{station_type}{nedap or fire stations}

\item{data}{all csv data of locations}
}
\value{
nedap or fire stations with same transform format data
}
\description{
transform
}
