% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_transform.R
\name{csv_tranform}
\alias{csv_tranform}
\title{csv_transform}
\usage{
csv_tranform(data, station_type)
}
\arguments{
\item{data}{all csv data of locations}

\item{station_type}{nedap or fire stations}
}
\value{
nedap or fire stations with same transform format data
}
\description{
csv_transform
}
\examples{
#temp1 <- csv_tranform(data = test, station_type = "nedap")
}
