% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_code.R
\name{dfi_get}
\alias{dfi_get}
\title{dfi_get}
\usage{
dfi_get(origin_data, adg_data)
}
\arguments{
\item{origin_data}{the results from import_csv}

\item{adg_data}{the results from adg_get}
}
\value{
dfi results
}
\description{
dfi_get
}
