% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_csv.R
\name{import_csv}
\alias{import_csv}
\title{import_csv}
\usage{
import_csv(file_list, package, ...)
}
\arguments{
\item{file_list}{full path list files}

\item{package}{from data.table, vroom or readr}

\item{...}{other parameters in three packages}
}
\value{
a data.table format
}
\description{
import_csv
}
\examples{
#test <- import_csv(file_list, package = "data.table", header = T, skip = 0)
#test <- import_csv(file = file_list, package = "data.table",header = F)
#test <- import_csv(file = file_list, package = "vroom", col_names = T)
#test <- import_csv(file = file_list, package = "readr", col_names = T, col_types = "dcddcddddc")
}
